+++ src/event/attachment_event.rs
@@ -41,12 +41,12 @@ impl SourceRef {
 #[derive(Serialize, Debug, Copy, Clone)]
 #[serde(rename_all = "camelCase")]
 pub struct Location {
-    line: usize,
-    column: usize,
+    line: u32,
+    column: u32,
 }
 
 impl Location {
-    pub fn new(line: usize, column: usize) -> Location {
+    pub fn new(line: u32, column: u32) -> Location {
         Location { line, column }
     }
 }
